requirements.txt
# ── 핵심 데이터 처리 ──────────────────────────────────────────
numpy>=2.0.0
pandas>=2.0.0
pyarrow>=14.0.0

# ── 최적화 ───────────────────────────────────────────────────
optuna>=3.5.0
scikit-learn>=1.3.0       # optuna fANOVA 파라미터 중요도

# ── 시각화 / 대시보드 ─────────────────────────────────────────
matplotlib>=3.8.0
streamlit>=1.37.0

# ── 데이터 수집 ───────────────────────────────────────────────
yfinance>=0.2.59
requests>=2.31.0
httpx>=0.27.0

# ── 환경변수 ─────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── PDF ──────────────────────────────────────────────────────
fpdf2>=2.7.0
pypdf>=4.0.0
pdfplumber>=0.10.0

# ── Alpaca 트레이딩 API ───────────────────────────────────────
alpaca-py>=0.20.0
